requirements.txt
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
pipenv==2022.8.15
